backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-jose[cryptography]
passlib[argon2]
argon2-cffi
python-multipart
sqladmin
itsdangerous
